fix: DRY delete article function
