+++ src/routes/article/[slug]/+page.svelte
@@ -60,7 +60,8 @@
 		</div>
 
 		<ArticleMetadata
-			id={article.author.id}
+			articleId={article.id}
+			userId={article.author.id}
 			nickname={article.author.nickname}
 			updated={article.updated}
 			isDeletable={isCurrentUserAuthor}
